need to fix sendTask
