+++ src/Components/Tasks/TasksForm.js
@@ -1,12 +1,16 @@
 //by kathleen tyner. This code presents the form the user will use to enter tasks into the dashboard.
 
 import { useState } from "react"
-import {Link} from "react-router-dom"
 import { useNavigate } from "react-router-dom"
 import { sendTask } from "./TasksApiManager"
-import "./tasks.css"
+import Calendar from 'react-calendar'
+
+import "./Tasks.css"
 
 export const TaskForm = () => {
+
+    const [date, onChange] = useState(new Date());
+
 
     const [task, update] = useState({
         task: "", 
@@ -26,7 +30,7 @@ export const TaskForm = () => {
         task: task.task,
         date: task.dateToComplete
     }
-    sendTask(taskSubmitAPI)
+   return sendTask()
     .then(( )=>  {
         navigate("/tasks") 
     
@@ -38,17 +42,17 @@ export const TaskForm = () => {
             <h2 className="taskForm__title">New Task</h2>
             <fieldset>
                 <div className="form-group">
-                    <label htmlFor="description">Task:</label>
+                    <label htmlFor="task">Task:</label>
                     <input
                         required autoFocus
                         type="text"
                         className="form-control"
-                        placeholder="Brief description of problem"
-                        value={task.task}
+                        placeholder="What do you need to do?"
+                        value={task?.task}
                         onChange={ 
                             (event) => {
-                            const copy = {...task} //make a copy of the ticket
-                            copy.description = event.target.value //look at the state of the ticket description
+                            const copy = {...task} 
+                            copy.task = event.target.value 
                             update(copy)
                         } 
                     }/>
@@ -56,17 +60,9 @@ export const TaskForm = () => {
             </fieldset>
             <fieldset>
                 <div className="form-group">
-                    <label htmlFor="name">Date</label>
-                    <input type="text"
-                        value={task.dateToComplete}
-                        onChange={
-                            (event) => {
-                                const copy = {...task}
-                                copy.dateToComplete= event.target.value //for checkboxes you do not need the value the check is a boolean
-                                update(copy)  // send the new state back
-
-                            }
-                        } />
+                    <label htmlFor="date">Date: </label><div>
+                    <Calendar onChange={onChange} value={task.dateToComplete} />
+                 </div>
                 </div>
             </fieldset>
             <button
